requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken
faiss-cpu
protobuf==3.19.0
langchain_experimental  # provide more information about this package
mysql-connector-python
pymysql
sentence-transformers
chromadb